Integration testing Complete!
The phase of integrating the smart contract codes into the hardhat suite for testing, to validate their definitions and structure has passed with 99.07% coverage of the code.
This was reached with manual creation of files, forcing the testing suite to test everything.
Some minor changes are required to the core contracts, and addition of pre-deployment protective layers before the final deployment.
However, they're considered complete at this point.